Extract detail component
